backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
pandas
numpy
django-environ
gunicorn